LU-2669 scripts: Change behavior of lustre_rmmod
The lustre_rmmod script was modified to take "-i" (inclusive) and
"-e" (exclusive) options along with a comma separated list modules,
rather than removing the ldiskfs and libcfs modules by default. For
example, to perform the old default functionality use:

    $ lustre_rmmod -i ldiskfs,libcfs

This will removed all depended modules to ldiskfs and libcfs, including
the ldiskfs and libcfs modules. If only the dependent modules should be
unloaded, excluding the ldiskfs and libcfs modules use:

    $ lustre_rmmod -e ldiskfs,libcfs

The test script, test-framework.sh, was updated to reflect the new
interface the this script. Also, the lnet init script now uses this
command to remove the lustre modules, rather than rolling its own
solution for removing the modules.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Id9732f399d1bc7d7d4f1ec4f684fa16cab35e786