LU-1199 obdclass: build and install llog_test.ko properly
There's a hack in lustre.spec.in to copy the llog_test.ko module into
/lib/modules.

It's only needed because that module is built unconditionally and
is expected to only be installed if --enable-tests is configured.

The right way to do this is to only build it at all if --enable-tests
is configured and installation comes automatically,

The conditional in the spec file for --enable-tests already exists,
so it will be packaged correctly without the hack.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Change-Id: I2705848e11a96ab6769ebba4481d2d0621f53d3e