LU-2904 llite: map ROOT FID to special local ino#
Since Lustre-2.4, non-IGIF FID for ROOT object will be
returned to client. For 64bit client, it converts such
ROOT FID into a local ino#, which is larger than 2^32.

For the case of re-exporting Lustre ROOT via NFS, it works
only when its ino# less than 2^32, which is NFS defect. So
if without proper handling, it cannot re-export the Lustre
ROOT via NFS. Similar issue exists all along on old Lustre
for re-exporting non-ROOT via NFS.

One solution is that map server-side ROOT FID to a special
local ino# - LL_ROOT_INO(12), which is an IGIF:

1) For other old IGIF objects returned by the server, they
   will not use the LL_ROOT_INO(12), because it is used by
   some server-side local object.

2) For other normal FID objects returned by the server, the
   existing converting mechanism should guarantee that the
   ino# converted from normal FID should NOT be conflict
   with IFIG. The "LL_ROOT_INO" will not make things worse.

If we only allow to re-export Lustre via NFS just against
the ROOT, then the client needs NOT to be mounted with the
option "-o 32bitapi".

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I6376f0afdd49cc0ffa11d896b813782ec1b2a471