LU-2904 fid: use IGIF for ROOT object
For the case of re-exporting Lustre ROOT via NFS, it works
only when its ino# less than 2^32, which is NFS defect. If
the MDT0 returns non-IGIF FID to client for the ROOT, then
the client (including the old Lustre-2.x client) will map
such non-IGIF FID into a local ino#, which is larger than
2^32, and then cause failure when re-exporting Lustre ROOT
via NFS.

On the other hand, for the MDT0 upgraded from old Lustre-2.x
device, it still returns IGIF for the old ROOT. That means
IGIF for ROOT object works under DNE mode. So we can specify
a special IGIF {FID_SEQ_IGIF, 1, 0} for the ROOT object. It
does not conflict with other IGIF object. Then FID_SEQ_ROOT
is useless any more.

If we only allow to re-export Lustre via NFS just against
the ROOT, then the client needs NOT to be mounted with the
option "-o 32bitapi".

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I6376f0afdd49cc0ffa11d896b813782ec1b2a471