LU-2893 lod: OST0 is unused when offset is specified
If a user specifies a non-zero OST index on which to start striping a
file, OST index 0 will never be used.  This is because the
lod_alloc_specific() function neglects to record used OST indexes as
it allocates OSTs. The array where the used indexes should have been
stored is initialized with zero's, so it always appears that index 0
is in use. Fix this by adding the missing statement to record used OST
indexes.

Signed-off-by: Ned Bass <bass6@llnl.gov>
Change-Id: I4b5526950ce8c3f9739838855d86c6fdb56bfe61