LU-2900 llite: Null deref in ll_fsync:mkdir on NFSmounted Lus
When a Lustre file system is mounted via NFS and a mkdir
operation is attempted, a null pointer dereference occurs
in ll_fsync.

With 2.x, Lustre added support for different VFS fsync APIs
that do not include a dentry parameter.

To make the logic the same in all cases, the old ll_fsync
interface was changed to pull the inode from the f_dentry
in the file parameter.

In some cases when using the old ll_fsync interface, the
caller does not set the file parameter resulting in a NULL
dereference. The fix to this is to restore the old logic in
those cases: when a dentry parameter is provided, get the
inode from that parameter rather than the file parameter.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I0f3ec27b84c6f4918d7bb26a54260b51313f70e9