LU-2907 build: Infiniband HW kernel modules of OFA builds not started
Nodes installed with rhel6 and ofa (external OFED) builds fail during
the node provisioning phase due to missing connectivity to the
Infiniband fabric. Reason is the fact that the HW kernel modules
mlx4_core, mlx4_en, mlx4_ib are not loaded (modprobe'd) during the
system boot phase.

For rhel5 an installation conflict of the startup script
'/etc/init.d/openibd' provided by the OFED kernel-ib RPM and a
distribution RPM (openibd) prohibited the installation of the
kernel-ib RPM. As a workaround the remove of the code sections
inside the SPEC file associated with the kernel-ib that provide the
packaging and configuration of the startup-script had been
implemented. This was accomplished when applying the ed script
'01-play-nice-with-RHEL5.ed' to the kernel-ib SPEC file.

The packaging structure of rhel6 has changed. The RPM opnenibd no
longer exist, therefore the startup of the HW kernel modules will
be missing for rhel6 and the symptom of missing connectivity
occurs.

The patch fixes the problem by searching (via regula expression) for
the canonical (distribution) target name within the name of ed
script and only apply the changes if the cananoical target matches
the ed script name.

ED scripts have to follow the naming convention:

<descriptive-name>-<canonical-target>.ed
or
<descriptive-name>-<canonical-target_1-...-<canonical-target_N>.ed

The string 'canonical-target' has to follow the convention used
for varialble CANONICAL_TARGET in script lbuild.
If a change should be applied to a single canonical target only the
first form has to be used.
If the change is valid form multiple targets the canonical targets
have to be specified as a dash separted list of targets.

To preserve the previous change for rhel5 the existing file
        'lustre/patches/ofed/01-play-nice-with-RHEL5.ed'
was renamed to
        'lustre/patches/ofed/01-play-nice-with-rhel5.ed'

Test-Parameters: nettypes=o2ib serverdistro=el6 clientibstack=ofa serveribstack=ofa

Signed-off-by: frank <Frank.Heckes@intel.com>
Change-Id: Ib25071e08553d28764e02ce50756deb91f757ed0