LU-2139 obdclass: Rename max_dirty_pages variable
This patch renames the "max_dirty_pages" global variable to
"max_pinned_pages". This is in anticipation for tracking the number of
outstanding unstable pages on the system, and limiting the sum of the
outstanding dirty and unstable pages by this value. Thus, the previous
name doesn't fit since it will be used to limit both dirty and unstable
pages. The mnemonic being both dirty and unstable pages are "pinned" in
memory, so this is the maximum number of lustre pages pinned in memory
at any point in time.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Ibed3865a80ef153855e9ca4b570837ac24a4df23