LU-2950 lnet: Add a mechanism to configure routes from a file
Created a bash script lustre_routes_config, which takes in a file
of the following format: <network> [hop] <gateway>[:<prioirty>]
It parses out each line of the file then invokes the following:
lctl --net <network> add_route <gateway> [hop [priority]]
This will configure the routes in the file in LNET.
It is possible to use the same script to remove the configured routes
by modifying the script to run: del_route.

This is an alternative implementation to Fujitsu implementationt to
modify the code to parse the lustre.conf file for a routes_config
parameter and open the file in kernel space to add the routes therein.

Signed-off-by: Amir Shehata <amir.shehata.whamcloud@gmail.com>
Change-Id: I57f81176c2d926fedefa8ea3be34586aa1ac9d76