LU-2800 llite: introduce local getname()
Commit 8e377d15 (v3.7) unexported getname and putname. Although
getname was re-exported, putname was not. All we really need is
something to to allocate a buffer and copy it from userspace so
let's just use our own version. The down size is that we lose
the audit support in the original getname.

Test-parameters: clientdistro=el6

Lustre-commit: 1bf4640f5a8e60d3f8d4676663a87c735ea57552
Lustre-change: http://review.whamcloud.com/8581

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: yang sheng <yang.sheng@intel.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Bobi Jam <bobijam@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: I02c28852ac5f4a5906ac43fac671dbacf3335acf