LU-2030 obdclass: Lustre returns EINTR during writes when SA_RESTART is set
When Lustre is in a read or write system call and receives a
SIGALRM, it currently returns EINTR at this location.  This is
problematic because it prevents the system call from being restarted
if SA_RESTART is set in the handler.

This patch changes this location to ERESTARTSYS.  This is a trial
to see if whether or not this is viable.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I26e24b8e8e325c5b0bd7d5d20fa97e2180c12263