LU-3044 llite: LSeek SEEK_CUR incorrect after write O_APPEND
When a file is opened with O_APPEND set and a write is done,
the file offset value immediately after the write is incorrect.
It is too much by approximately the length of the write.

This can be seen by doing lseek SEEK_CUR immediately after the
write.  This does not cause corruption on subsequent writes
because with O_APPEND VFS resets the file offset to EOF before
each write.

In ll_file_io_generic, the ppos, the file offset, is reset to
io->u.ci_wr.wr.crw_pos, which is incorrect when O_APPEND is
set.  When this is done, the existing value of ppos is correct.

This patch tests for O_APPEND and if it is set, does not set
ppos to io->u.ci_wr.wr.crw_pos.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I7c1ad10eefec44aae415b8cfce6b01bc9b39fc8f