FF-115 daos: DAOS event queue
All DAOS operations that cannot complete immediately return after
initiating the operation and post an event to an event queue to signal
completion asynchronously.
This patch is implementation of DAOS event queue (EQ) and event in
kernel space, userspace library will be implemented in upcoming patch.

This patch also includes implementation of DAOS handle table which is
based cfs_hash.

Signed-off-by: Liang Zhen <liang.zhen@intel.com>
Change-Id: I3e1185936fec711821854740db3f0d67ce5a2faa