LU-3059 lu: shrink lu_object by 16 bytes on x86_64
Remove the lo_depth member from struct lu_object.  This field is never
set and only read in lu_object_print().  Remove the lo_flags member.
This field was only used in lu_object_alloc() and can be replaced with
an on-stack mask to keep trace of which layers have been allocated.
Include the lo_dev_ref member if and only if it will be used, that is,
if and only if USE_LU_REF is defined.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I998cfda0234f4c418227c9ba640850a78ca7893e