LU-3044 llite: LSeek SEEK_CUR incorrect after O_APPEND write
When a file is opened for writing with O_APPEND set and data
is written, the offset in the file is incorrect after the write.
It is too large by approximately the size of the write.

This does not cause data corruption because VFS resets the
offset to end of file before each write when O_APPEND is set.

This is resolved by removing the change made for BUG:17711.
This is being pushed for testing - A commit message with
more detail is waiting on an update to LU-3044 from
Jinshan Xiong.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I0b262c77b6c54b864264a00314c98d178bb154c3