LU-2789 lod: synchronize access to ldo_stripe(nr)
Add a semaphore lod_stripe_sem to struct lod_object to synchronize
assignment to ldo_stripe(nr).  Acquire this semaphore when checking if
striping has been loaded for an object (as in the case of setattr()).
In lod_ah_init() use ldo_def_stripenr instead of ldo_stripenr to store
the stripe count hint and transfer it to ldo_stripenr later in
lod_qos_prep_create().  This removes LOD's dependence on
dt_write_lock() the next object to synchronize internally.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I253b4425fe22a46380603d6c17d81d4681fee20a