LU-1199 build: Rename ldiskfs rpm ldiskfs
The ldiskfs rpm package has been named lustre-ldiskfs for some
time, which seems to imply that it is created as part of the
lustre.spec file.  In reality, it was designed to be its own
stand-alone package with its own build system and separate spec
file.  This naming is less surprising.

Change-Id: I4dec26d15989fa644df155657c6bea7fdde2f12a
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>