LU-3093 lnet: Fix lnet selftest crash
When adding a group using lst add_group the group is added if and
only if at least one node is added successfully in the group.  There
are two cases that has to be handled:

1. An invlaid NID is provided, in which case the code still created
an empty group, but come time the group is used in a test, an assert
which checks that there is at least one node in the group is hit and
the entire operating system crashes.  Ideally there should be no
assertion on user input.

2. A valid NID is provided but the NID is not reachable.  In this cae
the node is added in UNKNOWN state, and test still doesn't run
successfully.

The fix is to ensure that a group has at least one acive node.
An addition of a non-empty group will not be supported.  Thus the
assert will never be hit.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I599c6351ac9aab4cc64a29f9f527cc1cfffaaecb