LU-1812 kernel: 3.0/SuSE and 3.6/FC18 server patches
This patch extends lustre server support to FC18 and
SLES11 SP2 with a ZFS. At this time ldiskfs is not
supported on FC18. Several outdated patches for SLES10
server server support also have been removed. The
jbd2-jcberr patch was not ported to these new platforms
for reasons described in LU-433

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I2a86d7ca4e068686372eb74f85b1443d24e245c4