LU-3152 utils: print lmm_oi correctly for old files
Lustre 2.1 through 2.3 stored the MDS fid.f_seq into lmm_object_seq
and fid.f_oid into lmm_object_id, but this was backward in Lustre
1.8 and earlier (storing the inode == IGIF f_seq into lmm_object_id,
and lmm_object_seq was always 0) and Lustre 2.4 and later (using this
space for an lu_fid).  These fields are used only in the client to
ensure inodes using DLM locks are unique, and a few error messages
in Lustre, and here when printing out the striping for userspace.

Handle this here transparently until it is fixed by LFSCK Phase 2,
since it will avoid potential future usage issues if it was done in
the kernel, since it might get confused if the version is ever used.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I2dea7eb1f9169f5b504f25d271ffeb48c7500c1e