LU-3041 lfsck: improve lfsck.sh on b1_8
The test directory of lfsck.sh contains some files referencing same
object, which could cause error when removing the directory on test
cleanup. Also, in lfsck.sh we shouldn't use debugfs to remove objects,
that'll cause quota usage inconsistence warning in e2fsck.

This patch is a backport of commit 622148 and 2df010, and it also
includes the patch of LU-3133.

Test-Parameters: testlist=lfsck

Signed-off-by: Liu Ying <emoly.liu@intel.com>
Change-Id: I6fd5db0c921e744f12a92676eb906d730b9c3d10