LU-3180 tests: handle FSCK_MAX_ERR properly
When the FSCK_MAX_ERR is 4 (File system errors left uncorrected),
we should not try to run the second lfsck, because the errors
could not be fixed due to read-only filesystem or some other
reason.

In the fix of LU-2780 (f9659fd87771122d01a4fafb09ba59c4cf0091ab),
the ./lustre/fid was changed to be an real inode, that'll beak
the is_empty_fs() of t-f. is_empty_fs() is changed accordingly.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I6216b5325ff7d9dc420d57a11f6d09f345bff96c