LU-3219 ost: extent lock is acquired during fiemap ioctl
Data corruption is possible if cp(coreutils) uses FIEMAP to obtain
data holes in copied files. FIEMAP ioctl was not prepared to deal with
regions of a file which have been written to, but which do not
actually have blocks assigned yet.

The way to be sure, that data flushed from clients is to have extent
lock on data range. This lock should be acquired it OST's FIEMAP
handler.

Xyratex-bug-id: MRP-1001
Signed-off-by: Artem Blagodarenko <artem_blagodarenko@xyratex.com>
Change-Id: I67fe05bac1ce75f98c3d3806a4d08f6cb02048e6