LU-1442 ost: extent lock is acquired during fiemap ioctl
Data corruption is possible if cp(coreutils) uses fiemap to obtain
data holes in copied files. fiemap ioctl was not prepared to deal with
regions of a file which have been written to, but which do not
actually have blocks assigned yet. The delayed allocation mechanism
used by most contemporary filesystems will create exactly that kind
of situation of data file, some blocks missing (i.e. zero'd) in
target.

The way to be sure, that data flushed from clients is to have extent
lock on data range. This lock should be acquired it OST's fiemap
handler.

Xyratex-bug-id: CLSTR-1160
Signed-off-by: Artem Blagodarenko <artem_blagodarenko@xyratex.com>
Change-Id: I67fe05bac1ce75f98c3d3806a4d08f6cb02048e6