LU-3223 utils: multiple args for path2fid/fid2path
Allow "lfs path2fid" to process multiple pathname arguments on
the command-line.  Like "grep" with multiple filenames, it will
print "filename: " at the beginning of the line when processing
multiple files.

Allow "lfd fid2path" to process multiple FID arguments on the
command-line.  All of the pathnames will be printed for each FID,
or can be limited to a single pathname per FID via "--link 0".

Add basic sanity tests for path2fid and fid2path, along with
tests of the multiple-argument options for each command.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I5c24b92af79cfeeafd4f5910e2f165dc3b500c1e